Store device info on init
