+++ lib/units/ios-device/plugins/group.js
@@ -7,7 +7,6 @@ import wireutil from '../../../wire/util.js'
 import * as grouputil from '../../../util/grouputil.js'
 import lifecycle from '../../../util/lifecycle.js'
 import * as dbapi from '../../../db/api.js'
-import request from 'request-promise'
 import * as iosutil from './util/iosutil.js'
 import * as apiutil from '../../../util/apiutil.js'
 import solo from './solo.js'
@@ -23,7 +22,6 @@ export default syrup.serial()
     .dependency(channels)
     .define((options, solo, router, push, sub, channels) => {
         const log = logger.createLogger('device:plugins:group')
-        const baseUrl = iosutil.getUri(options.wdaHost, options.wdaPort)
         let currentGroup = null
         let plugin = new events.EventEmitter()
         plugin.get = Promise.method(function() {
@@ -61,109 +59,6 @@ export default syrup.serial()
                         wireutil.global
                         , wireutil.envelope(new wire.JoinGroupMessage(options.serial, currentGroup, usage))
                     ])
-                    const handleRequest = (reqOptions) => {
-                        return new Promise((resolve, reject) => {
-                            request(reqOptions)
-                                .then((res) => {
-                                    resolve(res)
-                                })
-                                .catch((err) => {
-                                    reject(err)
-                                })
-                        })
-                    }
-                    dbapi.loadDeviceBySerial(options.serial).then((device) => {
-                        // No device size/type = First device session
-                        if (!device.display.width || !device.display.height || !device.deviceType) {
-                            // Get device type
-                            let deviceType
-                            handleRequest({
-                                method: 'GET'
-                                , uri: `${baseUrl}/wda/device/info`
-                                , json: true
-                            })
-                                .then((deviceInfo) => {
-                                    let deviceInfoModel = deviceInfo.value.model.toLowerCase()
-                                    let deviceInfoName = deviceInfo.value.name.toLowerCase()
-                                    if (deviceInfoModel.includes('tv') || deviceInfoName.includes('tv')) {
-                                        deviceType = 'Apple TV'
-                                    }
-                                    else {
-                                        deviceType = 'iPhone'
-                                    }
-                                    // Store device type
-                                    log.info('Storing device type value: ' + deviceType)
-                                    push.send([
-                                        wireutil.global
-                                        , wireutil.envelope(new wire.DeviceTypeMessage(options.serial, deviceType))
-                                    ])
-                                })
-                                .catch((err) => {
-                                    log.error('Error storing device type')
-                                    return lifecycle.fatal(err)
-                                })
-                            // Get device size
-                            handleRequest({
-                                method: 'POST'
-                                , uri: `${baseUrl}/session`
-                                , body: {capabilities: {}}
-                                , json: true,
-                            })
-                                .then((sessionResponse) => {
-                                    let sessionId = sessionResponse.sessionId
-                                    // Store device version
-                                    log.info('Storing device version')
-                                    push.send([
-                                        wireutil.global
-                                        , wireutil.envelope(new wire.SdkIosVersion(options.serial, sessionResponse.value.capabilities.sdkVersion))
-                                    ])
-                                    // Store battery info
-                                    if (deviceType !== 'Apple TV') {
-                                        handleRequest({
-                                            method: 'GET'
-                                            , uri: `${baseUrl}/session/${sessionId}/wda/batteryInfo`
-                                            , json: true,
-                                        })
-                                            .then((batteryInfoResponse) => {
-                                                let batteryState = iosutil.batteryState(batteryInfoResponse.value.state)
-                                                let batteryLevel = iosutil.batteryLevel(batteryInfoResponse.value.level)
-                                                push.send([
-                                                    wireutil.global
-                                                    , wireutil.envelope(new wire.BatteryEvent(options.serial, batteryState, 'good', 'usb', batteryLevel, 1, 0.0, 5))
-                                                ])
-                                            })
-                                            .catch((err) => log.error('Error storing battery', err))
-                                    }
-                                    // Store size info
-                                    return handleRequest({
-                                        method: 'GET'
-                                        , uri: `${baseUrl}/session/${sessionId}/window/size`
-                                        , json: true
-                                    }).then((firstSessionSize) => {
-                                        let deviceSize = firstSessionSize.value
-                                        let {width, height} = deviceSize
-                                        return handleRequest({
-                                            method: 'GET'
-                                            , uri: `${baseUrl}/session/${sessionId}/wda/screen`,
-                                        }).then((scaleResponse) => {
-                                            let parsedResponse = JSON.parse(scaleResponse)
-                                            let scale = parsedResponse.value.scale
-                                            height *= scale
-                                            width *= scale
-                                            log.info('Storing device size/scale')
-                                            push.send([
-                                                wireutil.global
-                                                , wireutil.envelope(new wire.SizeIosDevice(options.serial, height, width, scale))
-                                            ])
-                                        })
-                                    })
-                                })
-                                .catch((err) => {
-                                    log.error('Error storing device size/scale')
-                                    return lifecycle.fatal(err)
-                                })
-                        }
-                    })
                     plugin.emit('join', currentGroup)
                     return currentGroup
                 })
