+++ lib/units/ios-device/plugins/solo.js
@@ -39,7 +39,7 @@ export default syrup.serial()
                             })
                     })
                     .catch(err => {
-                        log.error('catch managerinfo', err)
+                        log.error('Error in manageDeviceInfo', err)
                         lifecycle.fatal(err)
                     })
             }
